bin/scout: starting the modularization of scout
scout's modules, command sets, and the executable itself is
now contained in bin/scout.

For better security, I'm starting to convert SNMP commands
to SSH. Created tests to ensure proper regex capturing of
data.
