+++ bin/scout/commands/scout_change_ap_ip
