+++ bin/scout/commands/scout_create_radius_ssid_5
