+++ bin/scout/commands/scout_create_ssid_24
