+++ bin/scout/commands/scout_delete_ssid_24
