+++ bin/scout/commands/scout_disable_ap_radius
