+++ bin/scout/commands/scout_do_wr
