+++ bin/scout/commands/scout_enable_ap_snmp
