+++ bin/scout/commands/scout_get_hostname
@@ -1,0 +1,1 @@
+
