+++ bin/scout/commands/scout_get_location
@@ -1,0 +1,3 @@
+enable
+{{ password }}
+show snmp location
