+++ bin/scout/commands/scout_get_users
@@ -1,0 +1,3 @@
+enable
+{{ password }}
+show users
