+++ bin/scout/commands/scout_reboot_ap
@@ -1,0 +1,5 @@
+enable
+{{ password }}
+reload
+y
+y
