+++ bin/scout/commands/scout_write_default
