+++ ci/tests/query-data-qa.py
@@ -29,9 +29,30 @@ SOFTWARE.
 # THIS IS ONLY A TEST FILE!
 
 import pathlib
+import re
 
-with pathlib.Path("ci/templates/getspeed").open(mode='r') as file:
+# Test querying AP model
+
+with pathlib.Path("ci/templates/version").open(mode='r') as file:
+    apModelRegex = re.compile(r'\w\w\w\-\w\w\w\w\w\w\w\w\-\w-\w\w')
+    getModelString = file.read().replace("\n", '')
+    getApModel = apModelRegex.search(getModelString).group(0)
+    print("--get-model OUTPUT RESULT: {}".format(getApModel))
+
+# Test querying MAC address
+
+with pathlib.Path("ci/templates/interface").open(mode='r') as file:
+    macAddrRegex = re.compile(r'\w\w\w\w.\w\w\w\w.\w\w\w\w')
+    getMacString = file.read().replace("\n", '')
+    getMacList = getMacString.split(",")
+    getMac = macAddrRegex.search(getMacList[2]).group(0)
+    print("--get-mac OUTPUT RESULT: {}".format(getMac))
+
+# Test querying AP speed
+
+with pathlib.Path("ci/templates/interface").open(mode='r') as file:
     getSpeedString = file.read().replace("\n", '')
     getSpeedList = getSpeedString.split(",")
     getSpeed = getSpeedList[9].strip("Mbps")
     print("--get-speed OUTPUT RESULT: {}".format(getSpeed))
+
